move version logic to its own internal package
